requirements.txt
einops==0.8.0
numpy==1.24.4
opencv_python==4.6.0.66
Pillow==11.0.0
scikit_image==0.17.2
skimage==0.0
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
